refactor: refactoring directories structure
